requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
py-cpuinfo==9.0.0
requests==2.28.1
tabulate==0.9.0
urllib3==1.26.12
wcwidth==0.2.5
yarg==0.1.9
